Fix uninitialized values and other small UT problems (#158)
* Fix strlen not finding a null in the log buffer

When deserializing a logger message containing a sized string argument,
the arg length is normally determined by finding a null character with
strlen(). But if the log buffer was too small to fit the argument or even
a (null-terminated) overflow marker, then no null gets stored and strlen()
will read beyond the end of the buffer and return incorrect length. This
tends to work though because the argument will be invalid and get ignored
anyway. Improve this by using strnlen() to cap the length to the amount of
bytes available in the buffer.

Change-Id: I41de8b6b2d66964e750b5ee23167b0b80b1695d5

* Fix uninitialized values and other small UT problems

Change-Id: I7a8b14d382987be6cd73d9d36e7e9c1db4702803

---------

Co-authored-by: Henri Bragge <henri.bragge@accenture.com>